knn/requirements.txt
streamlit>=1.49.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
scikit-image>=0.21.0
scipy>=1.10.0
